added authentication checking to the frontend

+++ frontend/src/components/WorkoutDetails.jsx
@@ -1,12 +1,19 @@
+import useAuthContext from "../hooks/useAuthContext"
 import useWorkoutsContext from "../hooks/useWorkoutsContext"
 import formatDistanceToNow from 'date-fns/formatDistanceToNow'
 
 export default function WorkoutDetails({ workout }) {
     const { dispatch } = useWorkoutsContext()
+    const { user } = useAuthContext()
 
     const handleClick = async (event) => {
+        if (!user) return
+
         const response = await fetch(`http://localhost:3000/api/workouts/${workout._id}`, {
-            method: 'DELETE'
+            method: 'DELETE',
+            headers: {
+                "Authorization": `Bearer ${user.token}`
+            }
         })
         const json = await response.json()
